requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
alembic=0.9.9=pypi_0
bcrypt=3.1.7=pypi_0
blinker=1.4=pypi_0
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py37_0
cffi=1.14.0=pypi_0
chardet=3.0.4=pypi_0
click=6.7=pypi_0
flask=1.0.2=pypi_0
flask-bcrypt=0.7.1=pypi_0
flask-dance=0.14.0=pypi_0
flask-debugtoolbar=0.10.1=pypi_0
flask-login=0.4.1=pypi_0
flask-migrate=2.1.1=pypi_0
flask-oauth=0.12=pypi_0
flask-oauthlib=0.9.4=pypi_0
flask-sqlalchemy=2.3.2=pypi_0
flask-wtf=0.14.2=pypi_0
httplib2=0.11.3=pypi_0
idna=2.6=pypi_0
itsdangerous=0.24=pypi_0
jinja2=2.10=pypi_0
lazy=1.3=pypi_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
mako=1.0.7=pypi_0
markupsafe=1.0=pypi_0
ncurses=6.2=h0a44026_0
oauth2=1.9.0.post1=pypi_0
oauthlib=2.0.7=pypi_0
openssl=1.1.1f=h1de35cc_0
pillow=7.1.2=pypi_0
pip=20.0.2=py37_1
pycparser=2.20=pypi_0
python=3.7.7=hc70fcce_0_cpython
python-dateutil=2.7.2=pypi_0
python-editor=1.0.3=pypi_0
readline=8.0=h1de35cc_0
requests=2.18.4=pypi_0
requests-oauthlib=0.8.0=pypi_0
setuptools=46.1.3=py37_0
six=1.12.0=pypi_0
sqlalchemy=1.2.6=pypi_0
sqlalchemy-utils=0.33.2=pypi_0
sqlite=3.31.1=ha441bb4_0
tk=8.6.8=ha441bb4_0
urllib3=1.22=pypi_0
urlobject=2.4.3=pypi_0
werkzeug=0.14.1=pypi_0
wheel=0.34.2=py37_0
wincertstore=0.2=pypi_0
wtforms=2.1=pypi_0
xz=5.2.4=h1de35cc_4
zlib=1.2.11=h1de35cc_3
